# src/backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PureDown API",
    "version": "0.2.0"
  },
  "paths": {
...
    "/api/analysis/preview": {
      "post": {
        "summary": "Preview Analysis",
        "operationId": "preview_analysis_api_analysis_preview_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalysisPreviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisPreview"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/analyses": {
      "post": {
        "summary": "Create Analysis",
        "operationId": "create_analysis_api_analyses_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAnalysisRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "List Analyses",
        "operationId": "list_analyses_api_analyses_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AnalysisResponse"
                  },
                  "title": "Response List Analyses Api Analyses Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/analyses/{analysis_id}": {
      "get": {
        "summary": "Get Analysis",
        "operationId": "get_analysis_api_analyses__analysis_id__get",
        "parameters": [
          {
            "name": "analysis_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Analysis Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/analyses/{analysis_id}/items/{item_id}/transcript": {
      "get": {
        "summary": "Get Transcript",
        "operationId": "get_transcript_api_analyses__analysis_id__items__item_id__transcript_get",
        "parameters": [
          {
            "name": "analysis_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Analysis Id"
            }
          },
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Item Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TranscriptSegment"
                  },
                  "title": "Response Get Transcript Api Analyses  Analysis Id  Items  Item Id  Transcript Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/analyses/{analysis_id}/items/{item_id}/retry": {
      "post": {
        "summary": "Retry Item",
        "operationId": "retry_item_api_analyses__analysis_id__items__item_id__retry_post",
        "parameters": [
          {
            "name": "analysis_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Analysis Id"
            }
          },
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Item Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/analyses/{analysis_id}/frames/{frame_id}": {
      "get": {
        "summary": "Get Frame",
        "operationId": "get_frame_api_analyses__analysis_id__frames__frame_id__get",
        "parameters": [
          {
            "name": "analysis_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Analysis Id"
            }
          },
          {
            "name": "frame_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Frame Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AnalysisItemResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "sourceUrl": {
            "type": "string",
            "title": "Sourceurl"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "position": {
            "type": "integer",
            "title": "Position"
          },
          "status": {
            "$ref": "#/components/schemas/AnalysisStatus"
          },
          "progress": {
            "type": "integer",
            "title": "Progress"
          },
          "source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "summary": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/VideoSummary"
              },
              {
                "type": "null"
              }
            ]
          },
          "frames": {
            "items": {
              "$ref": "#/components/schemas/FrameInfo"
            },
            "type": "array",
            "title": "Frames"
          }
        },
        "type": "object",
        "required": [
          "id",
          "sourceUrl",
          "title",
          "position",
          "status",
          "progress"
        ],
        "title": "AnalysisItemResponse"
      },
      "AnalysisPreview": {
        "properties": {
          "url": {
            "type": "string",
            "title": "Url"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "duration": {
            "type": "integer",
            "title": "Duration",
            "default": 0
          },
          "entries": {
            "items": {
              "$ref": "#/components/schemas/PreviewEntry"
            },
            "type": "array",
            "title": "Entries"
          },
          "totalDuration": {
            "type": "integer",
            "title": "Totalduration",
            "default": 0
          },
          "overLimit": {
            "type": "boolean",
            "title": "Overlimit",
            "default": false
          },
          "maxItems": {
            "type": "integer",
            "title": "Maxitems"
          }
        },
        "type": "object",
        "required": [
          "url",
          "title",
          "kind",
          "maxItems"
        ],
        "title": "AnalysisPreview"
      },
      "AnalysisPreviewRequest": {
        "properties": {
          "url": {
            "type": "string",
            "title": "Url"
          }
        },
        "type": "object",
        "required": [
          "url"
        ],
        "title": "AnalysisPreviewRequest"
      },
      "AnalysisResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "sourceUrl": {
            "type": "string",
            "title": "Sourceurl"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "status": {
            "$ref": "#/components/schemas/AnalysisStatus"
          },
          "progress": {
            "type": "integer",
            "title": "Progress"
          },
          "generatePlaylistOverview": {
            "type": "boolean",
            "title": "Generateplaylistoverview"
          },
          "playlistOverview": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PlaylistOverview"
              },
              {
                "type": "null"
              }
            ]
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/AnalysisItemResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "title": "Createdat"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "title": "Updatedat"
          }
        },
        "type": "object",
        "required": [
          "id",
          "sourceUrl",
          "title",
          "kind",
          "status",
          "progress",
          "generatePlaylistOverview",
          "createdAt",
          "updatedAt"
        ],
        "title": "AnalysisResponse"
      },
      "AnalysisStatus": {
        "type": "string",
        "enum": [
          "queued",
          "extracting",
          "transcribing",
          "summarizing",
          "completed",
          "partial",
          "failed"
        ],
        "title": "AnalysisStatus"
      },
...
      "CreateAnalysisRequest": {
        "properties": {
          "url": {
            "type": "string",
            "title": "Url"
          },
          "generatePlaylistOverview": {
            "type": "boolean",
            "title": "Generateplaylistoverview",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "url"
        ],
        "title": "CreateAnalysisRequest"
      },
...
      "FrameInfo": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "timestamp": {
            "type": "number",
            "title": "Timestamp"
          },
          "url": {
            "type": "string",
            "title": "Url"
          }
        },
        "type": "object",
        "required": [
          "id",
          "timestamp",
          "url"
        ],
        "title": "FrameInfo"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PlaylistOverview": {
        "properties": {
          "overview": {
            "type": "string",
            "title": "Overview"
          },
          "themes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Themes"
          },
          "sequence": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Sequence"
          },
          "omittedItems": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Omitteditems"
          }
        },
        "type": "object",
        "required": [
          "overview"
        ],
        "title": "PlaylistOverview"
      },
      "PreviewEntry": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "url": {
            "type": "string",
            "title": "Url"
          },
          "duration": {
            "type": "integer",
            "title": "Duration",
            "default": 0
          },
          "position": {
            "type": "integer",
            "title": "Position"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "url",
          "position"
        ],
        "title": "PreviewEntry"
      },
      "SummaryChapter": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "startSeconds": {
            "type": "number",
            "title": "Startseconds",
            "default": 0
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          },
          "keyPoints": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Keypoints"
          },
          "frameId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Frameid"
          }
        },
        "type": "object",
        "required": [
          "title",
          "summary"
        ],
        "title": "SummaryChapter"
      },
      "SummaryHighlight": {
        "properties": {
          "text": {
            "type": "string",
            "title": "Text"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags"
          }
        },
        "type": "object",
        "required": [
          "text"
        ],
        "title": "SummaryHighlight"
      },
      "TranscriptSegment": {
        "properties": {
          "startSeconds": {
            "type": "number",
            "title": "Startseconds"
          },
          "endSeconds": {
            "type": "number",
            "title": "Endseconds"
          },
          "text": {
            "type": "string",
            "title": "Text"
          }
        },
        "type": "object",
        "required": [
          "startSeconds",
          "endSeconds",
          "text"
        ],
        "title": "TranscriptSegment"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "VideoSummary": {
        "properties": {
          "language": {
            "type": "string",
            "title": "Language",
            "default": "zh"
          },
          "overview": {
            "type": "string",
            "title": "Overview"
          },
          "chapters": {
            "items": {
              "$ref": "#/components/schemas/SummaryChapter"
            },
            "type": "array",
            "title": "Chapters"
          },
          "highlights": {
            "items": {
              "$ref": "#/components/schemas/SummaryHighlight"
            },
            "type": "array",
            "title": "Highlights"
          },
          "questions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Questions"
          }
        },
        "type": "object",
        "required": [
          "overview"
        ],
        "title": "VideoSummary"
      }
    }
  }
}